Move Dotty test to slow tests in order to have proper mtags published
